refactor: split server-safe apiPublicRequest into api-public
