+++ src/lib/api.ts
@@ -1,8 +1,8 @@
-import { fetchApiJson } from '@/lib/api-core';
 import type { FetchApiJsonOptions } from '@/lib/api-core';
 import { fetchApiJsonWithSessionRecovery } from '@/lib/api-session-recovery';
 import { logger } from '@/lib/logger';
 
+export { apiPublicRequest } from '@/lib/api-public';
 export {
   ApiError,
   TELEGRAM_INIT_DATA_EXPIRED_CODE,
@@ -33,29 +33,6 @@ export async function apiRequest<TResponse = unknown, TBody = unknown>(
     });
   } catch (e) {
     logger.errorFromUnknown('api_request_failed', e, { endpointOrUrl, method });
-    throw e;
-  }
-}
-
-/**
- * Public API call. Omits cookies and disables session recovery.
- * Next.js cache behavior is configured explicitly at each server call site.
- */
-export async function apiPublicRequest<TResponse = unknown, TBody = unknown>(
-  endpointOrUrl: string,
-  method: HttpMethod,
-  data?: TBody,
-  init?: ApiRequestInit,
-): Promise<TResponse> {
-  try {
-    const headers = new Headers(init?.headers);
-    return await fetchApiJson<TResponse>(endpointOrUrl, method, data, {
-      ...init,
-      headers,
-      credentials: 'omit',
-    });
-  } catch (e) {
-    logger.errorFromUnknown('api_public_request_failed', e, { endpointOrUrl, method });
     throw e;
   }
 }
