+++ src/lib/i18n/translations.server.test.ts
@@ -4,7 +4,7 @@ const { getTranslationsApiPublicRequestMock } = vi.hoisted(() => ({
 
 vi.mock('server-only', () => ({}));
 
-vi.mock('@/lib/api', () => ({
+vi.mock('@/lib/api-public', () => ({
   apiPublicRequest: getTranslationsApiPublicRequestMock,
 }));
